create dedicated command infrastructure also fixes #15
