+++ src/chell.c
@@ -1,4 +1,5 @@
 #include "chell.h"
+#include "commands.h"
 
 int main()
 {
@@ -150,6 +151,7 @@ void executeCommand(char *commandString, struct executable *executables)
 
     pid_t processID;
     char *argv[PATH_MAX];
+    struct command* builtin;
 
     //Allocate the argv array
     for (int i = 0; i < PATH_MAX; i++)
@@ -183,7 +185,12 @@ void executeCommand(char *commandString, struct executable *executables)
         }
     }
 
-    if (programExists)
+    // check if the executed command is built into Chell
+    if ((builtin = is_builtin(argv[0])))
+    {
+        builtin->func(argv[1]);
+    }
+    else if (programExists)
     {
         processID = fork();
         if (processID == 0)
@@ -194,13 +201,6 @@ void executeCommand(char *commandString, struct executable *executables)
         }
         else
             waitpid(processID, 0, 0);
-    }
-    else if (strncmp(argv[0], "cd", 2) == 0)
-    {
-        if (argv[1] != NULL)
-            cd(argv[1]);
-        else
-            cd(getenv("HOME"));
     }
     else if (!programExists)
     {
@@ -215,22 +215,6 @@ void executeCommand(char *commandString, struct executable *executables)
 
 }
 
-void cd(char *path)
-{
-    int status = chdir(path);
-
-    if (status == 0)
-        setenv("PWD", getcwd(NULL, 4096), 1);
-    else if (status == -1)
-    {
-        if (errno == EACCES)
-            printf("%s: Permission denied.\n", path);
-        else if (errno == ENOENT)
-            printf("%s: Doesn't exist.\n", path);
-        else if (errno == ENOTDIR)
-            printf("%s: Not a directory.\n", path);
-    }
-}
 
 void sigintHandler(int signal_number)
 {
