+++ src/chell.h
@@ -1,12 +1,8 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
-#include <unistd.h>
 #include <sys/wait.h>
-#include <sys/types.h>
 #include <limits.h>
 #include <dirent.h>
-#include <errno.h>
 #include <signal.h> 
 
 #include "defs.h"
@@ -31,8 +27,6 @@ int splitString(char *split[], char *string, char *delim);
 int splitCommand(char *argv[], char *command);
 
 void executeCommand(char *commandString, struct executable *executables);
-
-void cd(char *path);
 
 void sigintHandler(int signal_number);
